Get rid of TODO in DatabaseConsistencyCheckModule
Method ExecuteAndGetTable new returns empty DataTable
if there is no output from the SQL Server. That means
there is no need for try/catch in
DatabaseConsistencyCheckModule.

Our implementation threw an exception if there was no
output from SQL Server. It was caused because there was
no check for table existence in DataSet.