Remove check for webparts in page template layout
It's not necessary. It was here because our consultants used
this query to see whether any pagetemplatelayout contains
repeater and than they looked if it is commented with HTML
comment (<-- <cms:cmsrepeater... > -->). Code Mirror
highlights this as a commented even it is actually not commented
because .net uses <% for comments.