Allow the viewing of Module comments separately from the running of the module
Currently, in order to see the details of what a module does, you have to Load the Module, executing its behavior. This is problematic because first-time users won't have the context of what the modules do in order to know which module they want to run. This commit adds a button "Show Comments" that displays the comments separately from the results. In order to do this without also displaying the "results", I had to remove the Results toggle from the header bar, and place it directly on the Toggle arrow and the Header text. No other changes to existing functionality were needed.