Fixed missing tables and classes error on Kentico 7
Added the tables  Analytics_ExitPages, Analytics_Index, CMS_Session, CMS_WebFarmServerTask and OM_ScoreContactRule to the whiteslist for tables for Kentico 7 and added the classes Ecommerce - SKU option and Ecommerce - Variant option to the whitelist for classes for Kentico 7. This fixes an issue when these are reported missing, as they are apparently "ghosts" of development that were never cleaned up, as noted in the issue #163 Kentico 7 sites report missing tables and classes.